+++ tests/testHelper.cpp
@@ -1,0 +1,5 @@
+#include <iostream>
+
+void cpp_console_out(const char *msg) {
+    std::cout << msg;
+}
